requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
unstructured
yfinance
requests
